replacd yandex id button in main screen
