+++ packages/client/src/pages/Home/components/LogoutPageMenu.tsx
@@ -1,3 +1,8 @@
+import { useAppDispatch } from '@/store/typedHooks'
+import {
+  oauthServiceFetch,
+  oauthRedirect,
+} from '@/store/slices/userSlice/thunks'
 import {
   List,
   ListItem,
@@ -16,6 +21,24 @@ const logoutMainPageMenu: { label: string; to: string }[] = [
 
 const LogoutPageMenu = ({ styles }: { styles: CSSModuleClasses }) => {
   const cx = classNames.bind(styles)
+  const dispatch = useAppDispatch()
+
+  const redirectToOauthYandexPage = (serviceId: string) => {
+    setTimeout(() => {
+      dispatch(oauthRedirect(serviceId))
+    }, 200)
+  }
+
+  const onAuthYandex = async () => {
+    try {
+      const serviceId = await dispatch(oauthServiceFetch()).unwrap()
+
+      redirectToOauthYandexPage(serviceId)
+    } catch (error) {
+      console.log('Error on OAuth fetch service id', error)
+    }
+  }
+
   return (
     <List>
       <Typography variant="h1" className={cx('menu__title')}>
@@ -33,6 +56,16 @@ const LogoutPageMenu = ({ styles }: { styles: CSSModuleClasses }) => {
           </Link>
         </ListItem>
       ))}
+      <ListItem className={cx('menu__item')}>
+        <div className={cx('menu__item-link')} onClick={onAuthYandex}>
+          <ListItemButton className={cx('link__button')}>
+            <ListItemText
+              primary="Войти с Yandex ID"
+              className={cx('link__button-label')}
+            />
+          </ListItemButton>
+        </div>
+      </ListItem>
     </List>
   )
 }
